add support for external checksums file
With this patch createrepo_c will process pre-built checksum lists taken
from the evironment variable `CREATEREPO_CHECKSUMS`.
If a package was found in the list the checksum is taken from the list
instead of creating the checksum of the package.
This can speed up createrepo_c significantly and reduce IO load in build
environments which provide those pre-built lists.